pip_requirements.txt
pytorch-lightning==1.7.7
transformers==4.23.1
datasets==2.6.1
evaluate==0.3.0
pyext==0.7
wandb==0.13.7
jupyter==1.0.0
jupyterlab==3.5.0
runai==0.4.1
hydra-core==1.2.0
pre-commit==2.20.0
hydra-colorlog==1.1.0
rich==12.6.0
numpy==1.23.0
pandas==1.5.0
matplotlib==3.5.2
jsonlines==3.1.0
sentencepiece==0.1.97
